% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsf.R
\name{factorialNK}
\alias{factorialNK}
\title{Calculate a part of the factorial, from i to j}
\usage{
factorialNK(i, j)
}
\arguments{
\item{i}{- integer, from}

\item{j}{- integer, to}
}
\value{
product from i to j
}
\description{
Calculate a part of the factorial, from i to j
}
\examples{
factorialNK(3,7)

}
