% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsf.R
\name{FScalarProduct}
\alias{FScalarProduct}
\title{Calculate a scalar product of the spherical harmonics with indexes i and j}
\usage{
FScalarProduct(i, j, where = WHOLE_SPHERE)
}
\arguments{
\item{i}{- index, integer}

\item{j}{- index, integer}

\item{where}{- flag, where the value must be calculated, 1 - whole sphere, 0 - hemisphere}
}
\description{
Calculate a scalar product of the spherical harmonics with indexes i and j
}
