% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsf.R
\name{GetJbyNKP}
\alias{GetJbyNKP}
\title{Calculation combined index j from indexes n, k, p}
\usage{
GetJbyNKP(n, k, p)
}
\arguments{
\item{n}{- index, integer}

\item{k}{- index, integer}

\item{p}{- index, integer}
}
\value{
combined index j, integer
}
\description{
Calculation combined index j from indexes n, k, p
}
