% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsf.R
\name{GetNKPbyJ}
\alias{GetNKPbyJ}
\title{Calculation indexes n, k, p from combined index j}
\usage{
GetNKPbyJ(j)
}
\arguments{
\item{combined}{index j, integer}
}
\value{
list with 3 indexes: n, k, p (integer)
}
\description{
Calculation indexes n, k, p from combined index j
}
