% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsf.R
\name{GetSphFuncK_matrix}
\alias{GetSphFuncK_matrix}
\title{Calculate a spherical harmonic within a range from J = 0 to J = j given and with given (l, b)}
\usage{
GetSphFuncK_matrix(j, l, b)
}
\arguments{
\item{j}{- maximum combined index}

\item{l}{- vector of longitudes}

\item{b}{- vector of latitudes}
}
\value{
matrix of calculated spherical harmonics
}
\description{
Calculate a spherical harmonic within a range from J = 0 to J = j given and with given (l, b)
}
