% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsf.R
\name{LegendrePolinom}
\alias{LegendrePolinom}
\title{Calculate of the associated Legendre polynom in b}
\usage{
LegendrePolinom(n, k, b)
}
\arguments{
\item{n}{- Legendre polinom index, integer}

\item{k}{- Legendre polinom index, integer}

\item{b}{- point where polynom value must be calculated, real number}
}
\value{
value of associated Legendre polinom with indexes n and k in point b
}
\description{
Calculate of the associated Legendre polynom in b
}
\examples{
LegendrePolinom(3, 1, 0.5)

}
