% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsf.R
\name{SphFuncK_NKP}
\alias{SphFuncK_NKP}
\title{Calculation of the spherical harmonic value with indexes n, k and p in the point (l, b)}
\usage{
SphFuncK_NKP(n, k, p, l, b, where = WHOLE_SPHERE)
}
\arguments{
\item{n}{- index, integer}

\item{k}{- index, integer}

\item{p}{- index, integer, can be 0 or 1}

\item{l}{- longitude, real}

\item{b}{- latitude, real}

\item{where}{- flag, where the value must be calculated, 1 - whole sphere, 0 - hemisphere}
}
\value{
spherical harmonic value with indexes n, k and p in the point(l, b)
}
\description{
Calculation of the spherical harmonic value with indexes n, k and p in the point (l, b)
}
