% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsf.R
\name{SphNormaR}
\alias{SphNormaR}
\title{Calculation of the normalization factor of the spherical harmonic with indexes n and k}
\usage{
SphNormaR(n, k, where = WHOLE_SPHERE)
}
\arguments{
\item{n}{- index, integer}

\item{k}{- index, integer}

\item{where}{- flag, where the value must be calculated, 1 - whole sphere, 0 - hemisphere}
}
\value{
normalization factor value, real
}
\description{
Calculation of the normalization factor of the spherical harmonic with indexes n and k
}
