% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsf.R
\name{VSphFuncSB_J}
\alias{VSphFuncSB_J}
\title{Calculate the value of the sectorial vector spherical harmonic in latitude with index j in the point (l, b)}
\usage{
VSphFuncSB_J(j, l, b, where = WHOLE_SPHERE)
}
\arguments{
\item{j}{- combined index, integer}

\item{l}{- longitude, real}

\item{b}{- latitude, real}

\item{where}{- flag, where the value must be calculated, 1 - whole sphere, 0 - hemisphere}
}
\value{
sectorial vector spherical harmonic value with index j in the point(l, b) in latitude
}
\description{
Calculate the value of the sectorial vector spherical harmonic in latitude with index j in the point (l, b)
}
