% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsf.R
\name{VSphFuncSB_NKP}
\alias{VSphFuncSB_NKP}
\title{Calculate the value of the sectorial vector spherical harmonic in latitude with indexes n, k and p in the point (l, b)}
\usage{
VSphFuncSB_NKP(n, k, p, l, b, where = WHOLE_SPHERE)
}
\arguments{
\item{n}{- index, integer}

\item{k}{- index, integer}

\item{p}{- index, integer, can be 0 or 1}

\item{l}{- longitude, real}

\item{b}{- latitude, real}

\item{where}{- flag, where the value must be calculated, 1 - whole sphere, 0 - hemisphere}
}
\value{
sectorial vector spherical harmonic value with index j in the point(l, b) in latitude
}
\description{
Calculate the value of the sectorial vector spherical harmonic in latitude with indexes n, k and p in the point (l, b)
}
