% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsf.R
\name{VSphFuncTL_J}
\alias{VSphFuncTL_J}
\title{Calculate the value of the tesseral vector spherical harmonic in longitude with index j in the point (l, b)}
\usage{
VSphFuncTL_J(j, l, b, where = WHOLE_SPHERE)
}
\arguments{
\item{j}{- combined index, integer}

\item{l}{- longitude, real}

\item{b}{- latitude, real}

\item{where}{- flag, where the value must be calculated, 1 - whole sphere, 0 - hemisphere}
}
\value{
tesseral vector spherical harmonic value with index j in the point(l, b)
}
\description{
Calculate the value of the tesseral vector spherical harmonic in longitude with index j in the point (l, b)
}
