% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsf.R
\name{VSphScalarProduct}
\alias{VSphScalarProduct}
\title{Calculation a of the scalar product of the vector sperical harmonics with indexes i and j}
\usage{
VSphScalarProduct(i, j, what, where = WHOLE_SPHERE, FUN = function(l, b, i)
  return(0))
}
\arguments{
\item{i}{- index, integer}

\item{j}{- index, integer}

\item{what}{- what kind of function must be used}

\item{where}{- flag, where the value must be calculated, 1 - whole sphere, 0 - hemisphere}
}
\description{
Calculation a of the scalar product of the vector sperical harmonics with indexes i and j
}
