% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsf.R
\name{VSphScalarProductField}
\alias{VSphScalarProductField}
\title{Calculation a of the scalar product of the vector sperical harmonics with given field}
\usage{
VSphScalarProductField(i, what, field)
}
\arguments{
\item{i}{- index, integer}

\item{what}{- what kind of function must be used}

\item{field}{-}
}
\value{
-
}
\description{
Calculation a of the scalar product of the vector sperical harmonics with given field
}
